feat: 메타데이터 구현 및 배포서버 테스트
